Split preRenderService from renderService #178
